app/requirements.txt
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.0
h11==0.14.0
idna==3.4
paho-mqtt==1.6.1
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
redis==5.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2